refactored to use base class for getting and setting promises
