+++ client/src/luigi-client.js
@@ -1,4 +1,5 @@
-import { uxManager } from './uxManager';
+import { LuigiClientLinkManager } from './linkManager';
+import { LuigiClientUxManager } from './uxManager';
 
 let luigiInitialized = false;
 const defaultContextKeys = ['context', 'internal', 'nodeParams', 'pathParams'];
@@ -10,7 +11,7 @@ let currentContext = defaultContextKeys.reduce(function(acc, key) {
 const _onContextUpdatedFns = {};
 const _onInitFns = {};
 let authData = {};
-const pathExistsPromises = {};
+const linkManager = new LuigiClientLinkManager({ currentContext });
 
 /**
  * Creates a random Id
@@ -96,31 +97,6 @@ const luigiClientInit = () => {
       );
     } else if ('luigi.auth.tokenIssued' === e.data.msg) {
       setAuthData(e.data.authData);
-    }
-
-    if ('luigi.navigation.pathExists.answer' === e.data.msg) {
-      var data = e.data.data;
-      pathExistsPromises[data.correlationId].resolveFn(data.pathExists);
-      delete pathExistsPromises[data.correlationId];
-    }
-
-    if ('luigi.ux.confirmationModal.hide' === e.data.msg) {
-      const data = e.data.data;
-      const promise = uxManager().getPromise('confirmationModal');
-      if (promise) {
-        data.confirmed ? promise.resolveFn() : promise.rejectFn();
-        uxManager().setPromise('confirmationModal', undefined);
-      }
-    }
-
-    if ('luigi.ux.alert.hide' === e.data.msg) {
-      const { id } = e.data;
-      const alerts = uxManager().getPromise('alerts');
-      if (id && alerts[id]) {
-        alerts[id].resolveFn(id);
-        delete alerts[id];
-        uxManager().setPromise('alerts', alerts);
-      }
     }
   });
 
@@ -217,200 +193,12 @@ const LuigiClient = {
    */
   getPathParams: () => currentContext.pathParams,
   /**
-     * The Link Manager allows you to navigate to another route. Use it instead of an internal router to:
-      - Route inside micro front-ends.
-      - Reflect the route.
-      - Keep the navigation state in Luigi.
-     */
-  /** @name linkManager */
-  linkManager: function linkManager() {
-    var options = {
-      preserveView: false,
-      nodeParams: {},
-      errorSkipNavigation: false,
-      fromContext: null,
-      fromClosestContext: false,
-      relative: false,
-      link: ''
-    };
-
-    return {
-      /** @lends linkManager */
-      /**
-       * Navigates to the given path in the application hosted by Luigi. It contains either a full absolute path or a relative path without a leading slash that uses the active route as a base. This is the standard navigation.
-       * @param {string} path path to be navigated to
-       * @param {string} [sessionId] current Luigi **sessionId**
-       * @param {boolean} [preserveView] preserve a view by setting it to `true`. It keeps the current view opened in the background and opens the new route in a new frame. Use the {@link #goBack goBack()} function to navigate back. You can use this feature across different levels. Preserved views are discarded as soon as the standard {@link #navigate navigate()} function is used instead of {@link #goBack goBack()}
-       * @param {Object} [modalSettings] opens a view in a modal. Use these settings to configure the modal's title and size
-       * @param {string} modalSettings.title modal title. By default, it is the node label. If there is no label, it is left empty
-       * @param {('l'|'m'|'s')} [modalSettings.size="l"] size of the modal
-       * @example
-       * LuigiClient.linkManager().navigate('/overview')
-       * LuigiClient.linkManager().navigate('users/groups/stakeholders')
-       * LuigiClient.linkManager().navigate('/settings', null, true) // preserve view
-       */
-      navigate: function navigate(
-        path,
-        sessionId,
-        preserveView,
-        modalSettings
-      ) {
-        if (options.errorSkipNavigation) {
-          options.errorSkipNavigation = false;
-          return;
-        }
-        options.preserveView = preserveView;
-        var relativePath = path[0] !== '/';
-        var navigationOpenMsg = {
-          msg: 'luigi.navigation.open',
-          sessionId: sessionId,
-          params: Object.assign(options, {
-            link: path,
-            relative: relativePath,
-            modal: modalSettings
-          })
-        };
-        window.parent.postMessage(navigationOpenMsg, '*');
-      },
-      /**
-       * Opens a view in a modal. You can specify the modal's title and size. If you don't specify the title, it is the node label. If there is no node label, the title remains empty.  The default size of the modal is `l`, which means 80%. You can also use `m` (60%) and `s` (40%) to set the modal size. Optionally, use it in combination with any of the navigation functions.
-       * @param {string} path navigation path
-       * @param {Object} [modalSettings] opens a view in a modal. Use these settings to configure the modal's title and size
-       * @param {string} modalSettings.title modal title. By default, it is the node label. If there is no label, it is left empty
-       * @param {('l'|'m'|'s')} [modalSettings.size="l"] size of the modal
-       * @example
-       * LuigiClient.linkManager().openAsModal('projects/pr1/users', {title:'Users', size:'m'});
-       */
-      openAsModal: function(path, modalSettings) {
-        this.navigate(path, 0, true, modalSettings || {});
-      },
-      /**
-       * Sets the current navigation context to that of a specific parent node which has the {@link navigation-configuration.md navigationContext} field declared in the navigation configuration. This navigation context is then used by the `navigate` function.
-       * @param {string} navigationContext
-       * @returns {linkManager} link manager instance
-       * @example
-       * LuigiClient.linkManager().fromContext('project').navigate('/settings')
-       */
-      fromContext: function fromContext(navigationContext) {
-        var navigationContextInParent =
-          currentContext.context.parentNavigationContexts.indexOf(
-            navigationContext
-          ) !== -1;
-        if (navigationContextInParent) {
-          options.fromContext = navigationContext;
-        } else {
-          options.errorSkipNavigation = true;
-          console.error(
-            'Navigation not possible, navigationContext ' +
-              navigationContext +
-              ' not found.'
-          );
-        }
-        return this;
-      },
-
-      /**
-       * Sets the current navigation context which is then used by the `navigate` function. This has to be a parent navigation context, it is not possible to use the child navigation contexts.
-       * @returns {linkManager} link manager instance
-       * @example
-       * LuigiClient.linkManager().fromClosestContext().navigate('/users/groups/stakeholders')
-       */
-      fromClosestContext: function fromClosestContext() {
-        var hasParentNavigationContext =
-          currentContext.context.parentNavigationContexts.length > 0;
-        if (hasParentNavigationContext) {
-          options.fromContext = null;
-          options.fromClosestContext = true;
-        } else {
-          console.error(
-            'Navigation not possible, no parent navigationContext found.'
-          );
-        }
-        return this;
-      },
-
-      /**
-       * Sends node parameters to the route. The parameters are used by the `navigate` function. Use it optionally in combination with any of the navigation functions and receive it as part of the context object in Luigi Client.
-       * @param {Object} nodeParams
-       * @returns {linkManager} link manager instance
-       * @example
-       * LuigiClient.linkManager.withParams({foo: "bar"}).navigate("path")
-       *
-       * // Can be chained with context setting functions such as:
-       * LuigiClient.linkManager.fromContext("currentTeam").withParams({foo: "bar"}).navigate("path")
-       */
-      withParams: function withParams(nodeParams) {
-        if (nodeParams) {
-          Object.assign(options.nodeParams, nodeParams);
-        }
-        return this;
-      },
-
-      /** @lends linkManager */
-      /**
-       * Checks if the path you can navigate to exists in the main application. For example, you can use this helper method conditionally to display a DOM element like a button.
-       * @param {string} path path which existence you want to check
-       * @returns {promise} a promise which resolves to a Boolean variable specifying whether the path exists or not
-       * @example
-       *  let pathExists;
-       *  this.luigiClient
-       *  .linkManager()
-       *  .pathExists('projects/pr2')
-       *  .then(
-       *    (pathExists) => {  }
-       *  );
-       */
-      pathExists: function pathExists(path) {
-        var currentId = Date.now();
-        pathExistsPromises[currentId] = {
-          resolveFn: function() {},
-          then: function(resolveFn) {
-            this.resolveFn = resolveFn;
-          }
-        };
-        var pathExistsMsg = {
-          msg: 'luigi.navigation.pathExists',
-          data: {
-            id: currentId,
-            link: path,
-            relative: path[0] !== '/'
-          }
-        };
-        window.parent.postMessage(pathExistsMsg, '*');
-        return pathExistsPromises[currentId];
-      },
-
-      /**
-       * Checks if there is one or more preserved views. You can use it to show a **back** button.
-       * @returns {boolean} indicating if there is a preserved view you can return to
-       */
-      hasBack: function hasBack() {
-        return (
-          !!currentContext.internal.modal ||
-          currentContext.internal.viewStackSize !== 0
-        );
-      },
-
-      /**
-       * Discards the active view and navigates back to the last visited view (preserved view), if a preserved view was set before.
-       * @param {any} goBackValue data that is passed in the `goBackContext` field to the last visited view
-       * @example
-       * LuigiClient.linkManager().goBack({ foo: 'bar' });
-       * LuigiClient.linkManager().goBack(true);
-       */
-      goBack: function goBack(goBackValue) {
-        if (this.hasBack()) {
-          window.parent.postMessage(
-            {
-              msg: 'luigi.navigation.back',
-              goBackContext: goBackValue && JSON.stringify(goBackValue)
-            },
-            '*'
-          );
-        }
-      }
-    };
+   * @private
+   */
+  linkManager: () => {
+    linkManager.setCurrentContext(currentContext);
+    return linkManager;
   },
-  uxManager
+  uxManager: () => new LuigiClientUxManager()
 };
 export default LuigiClient;
